Move SpawnManager to world.scripts
