+++ scripts/__init__.py
@@ -1,5 +1,5 @@
 from .builder_autosave import BuilderAutosave
 from .sated_decay import SatedDecayScript
 from .global_npc_ai import GlobalNPCAI
-from .spawn_manager import SpawnManager
+from world.scripts.spawn import SpawnManager
