+++ world/scripts/__init__.py
@@ -1,0 +1,3 @@
+from .spawn import SpawnManager
+
+__all__ = ["SpawnManager"]
